generatorSinglePage/requirements.txt
asgiref==3.2.7
Django==3.0.7
django-crispy-forms==1.9.1
django-hitcount==1.3.2
gitdb==4.0.5
GitPython==3.1.3
Pillow==7.1.2
pkg-resources==0.0.0
pycodestyle==2.6.0
pytz==2020.1
smmap==3.0.4
sqlparse==0.3.1
toml==0.10.1
